+++ vscode/src/testutils/mocks.ts
@@ -10,7 +10,7 @@ import type {
 } from 'vscode'
 
 import {
-    type Configuration,
+    type ClientConfiguration,
     type FeatureFlag,
     FeatureFlagProvider,
     OLLAMA_DEFAULT_URL,
@@ -935,4 +935,4 @@ export const DEFAULT_VSCODE_SETTINGS = {
     autocompleteExperimentalHotStreakAndSmartThrottle: false,
     testingModelConfig: undefined,
     experimentalGuardrailsTimeoutSeconds: undefined,
-} satisfies Configuration
+} satisfies ClientConfiguration